refactor: Modularization of the schemas files
Refs #FLU-213
