+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './controllers/index.js';
 export * from './services/index.js';
-export * from './schemas.js';
 export * from './types/index.js';
+export * from './prompts/index.js';
