+++ src/prompts/index.ts
@@ -1,0 +1,1 @@
+export * from './operation-descriptions.js';
