+++ src/prompts/operation-descriptions.ts
@@ -1,126 +1,4 @@
-import { z } from 'zod';
-
-/**
- * Fluora operation types - following fluora-mcp naming standards
- */
-export enum FluoraOperation {
-  // Discovery Operations
-  SEARCH_FLUORA = 'search-fluora',
-  LIST_SERVERS = 'list-servers',
-  GET_SERVER_INFO = 'get-server-info',
-
-  // Monetization Operations (from MonetizedMCPServer)
-  PRICE_LISTING = 'price-listing',
-  PAYMENT_METHODS = 'payment-methods',
-  MAKE_PURCHASE = 'make-purchase',
-
-  // Tool Execution Operations
-  CALL_SERVER_TOOL = 'call-server-tool',
-  LIST_TOOLS = 'list-tools',
-
-  // Management Operations
-  GET_PURCHASE_HISTORY = 'get-purchase-history',
-  VALIDATE_PAYMENT = 'validate-payment',
-}
-
-/**
- * Payment methods supported by the system (from monetized-mcp-sdk)
- */
-export enum PaymentMethods {
-  USDC_BASE_SEPOLIA = 'USDC_BASE_SEPOLIA',
-  USDC_BASE_MAINNET = 'USDC_BASE_MAINNET',
-}
-
-export const SearchServerSchema = z.object({
-  operation: z.literal(FluoraOperation.SEARCH_FLUORA),
-  name: z.string().optional(),
-});
-
-export const ListServersSchema = z.object({
-  operation: z.literal(FluoraOperation.LIST_SERVERS),
-  category: z.string().optional(),
-});
-
-export const GetServerInfoSchema = z.object({
-  operation: z.literal(FluoraOperation.GET_SERVER_INFO),
-  serverId: z.string(),
-});
-
-export const PriceListingSchema = z.object({
-  operation: z.literal(FluoraOperation.PRICE_LISTING),
-  serverId: z.string(),
-});
-
-export const PaymentMethodsSchema = z.object({
-  operation: z.literal(FluoraOperation.PAYMENT_METHODS),
-  serverId: z.string(),
-});
-
-export const MakePurchaseSchema = z.object({
-  operation: z.literal(FluoraOperation.MAKE_PURCHASE),
-  serverId: z.string(),
-  itemPrice: z.string().regex(/^\d*\.?\d+$/, 'Invalid price format'),
-  serverWalletAddress: z
-    .string()
-    .regex(/^0x[a-fA-F0-9]{40}$/, 'Invalid wallet address'),
-});
-
-export const CallServerToolSchema = z.object({
-  operation: z.literal(FluoraOperation.CALL_SERVER_TOOL),
-  serverId: z.string(),
-  toolName: z.string(),
-  args: z.record(z.any()),
-});
-
-export const ListToolsSchema = z.object({
-  operation: z.literal(FluoraOperation.LIST_TOOLS),
-  serverId: z.string(),
-});
-
-export const GetPurchaseHistorySchema = z.object({
-  operation: z.literal(FluoraOperation.GET_PURCHASE_HISTORY),
-  userAddress: z
-    .string()
-    .regex(/^0x[a-fA-F0-9]{40}$/, 'Invalid address')
-    .optional(),
-});
-
-export const ValidatePaymentSchema = z.object({
-  operation: z.literal(FluoraOperation.VALIDATE_PAYMENT),
-  transactionHash: z
-    .string()
-    .regex(/^0x[a-fA-F0-9]{64}$/, 'Invalid transaction hash'),
-  paymentMethod: z.nativeEnum(PaymentMethods),
-});
-
-/**
- * Discriminated union created from individual schema objects
- */
-export const fluoraOperationSchema = z.discriminatedUnion('operation', [
-  SearchServerSchema,
-  ListServersSchema,
-  GetServerInfoSchema,
-  PriceListingSchema,
-  PaymentMethodsSchema,
-  MakePurchaseSchema,
-  CallServerToolSchema,
-  ListToolsSchema,
-  GetPurchaseHistorySchema,
-  ValidatePaymentSchema,
-]);
-
-/**
- * Create a Zod enum schema for FluoraOperation
- */
-export const FluoraOperationSchema = z.nativeEnum(FluoraOperation);
-
-/**
- * Create a map schema where keys are FluoraOperation and values are descriptions
- */
-export const OperationDescriptionSchema = z.record(
-  FluoraOperationSchema,
-  z.string()
-);
+import { FluoraOperation } from '../types/operations.js';
 
 /**
  * Operation descriptions for each Fluora operation
@@ -223,26 +101,10 @@ A successful response will return a JSON like this:
             "type": "string"
           },
           "params": {
-            "type": "object",
-            "additionalProperties": {}
-          },
-          "signedTransaction": {
-            "type": "string"
-          },
-          "paymentMethod": {
-            "type": "string",
-            "enum": [
-              "USDC_BASE_SEPOLIA",
-              "USDC_BASE_MAINNET"
-            ]
+            "type": "object"
           }
         },
-        "required": [
-          "itemId",
-          "params",
-          "signedTransaction",
-          "paymentMethod"
-        ],
+        "required": ["itemId"],
         "additionalProperties": false,
         "$schema": "http://json-schema.org/draft-07/schema#"
       }
@@ -330,50 +192,3 @@ A failure response will return an error message with details.
   ],
 ]);
 
-/**
- * Precheck success result schema
- */
-export const precheckSuccessSchema = z.object({
-  operationValid: z.boolean(),
-  serverAccessible: z.boolean().optional(),
-  estimatedCost: z.string().optional(),
-  paymentMethodValid: z.boolean().optional(),
-  walletAddressValid: z.boolean().optional(),
-});
-
-/**
- * Precheck failure result schema
- */
-export const precheckFailSchema = z.object({
-  error: z.string(),
-  operationType: z.string().optional(),
-});
-
-/**
- * Execute success result schema
- */
-export const executeSuccessSchema = z.object({
-  operation: z.nativeEnum(FluoraOperation),
-  result: z.any(),
-  transactionHash: z.string().optional(),
-  timestamp: z.number(),
-  serverId: z.string().optional(),
-  cost: z.string().optional(),
-});
-
-/**
- * Execute failure result schema
- */
-export const executeFailSchema = z.object({
-  error: z.string(),
-  operation: z.nativeEnum(FluoraOperation),
-  timestamp: z.number(),
-});
-
-// Type exports
-export type ToolParams = z.infer<typeof fluoraOperationSchema>;
-export type PrecheckSuccess = z.infer<typeof precheckSuccessSchema>;
-export type PrecheckFail = z.infer<typeof precheckFailSchema>;
-export type ExecuteSuccess = z.infer<typeof executeSuccessSchema>;
-export type ExecuteFail = z.infer<typeof executeFailSchema>;
-export type OperationDescriptions = z.infer<typeof OperationDescriptionSchema>;
