+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './mcp.js';
 export * from './discovery.js';
+export * from './operations.js';
